chore :modifying collection

+++ src/com/java/advanced/collectionsFramework/AllCollection.java
@@ -7,7 +7,7 @@ public class AllCollection {
     public static void main(String[] args) {
         Stack<Integer>stack=new Stack<>();
         for(int i = 0;i<10;i++) {
-            stack.push(i);
+            stack.push((Integer) i);
             System.out.println(stack.peek());
         }
         System.out.println(stack);
